+++ src/main/java/br/com/henrique/repository/RotaEntregaRepository.java
@@ -1,11 +1,13 @@
 package br.com.henrique.repository;
 
+import java.util.Optional;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import br.com.henrique.model.RotaEntrega;
 
 @Repository
-public interface RotaEntregaRepository extends JpaRepository<RotaEntrega, Long> {
+public interface RotaEntregaRepository extends JpaRepository<RotaEntrega, Integer> {
 
 }
